# Conflicts:
#	server/src/main/antlr3/lsfusion/server/language/LsfLogics.g
#	server/src/main/java/lsfusion/server/language/ScriptingLogicsModule.java
#	server/src/main/java/lsfusion/server/logics/LogicsModule.java
#	server/src/main/java/lsfusion/server/logics/action/flow/JoinAction.java
#	server/src/main/resources/ServerResourceBundle.properties
#	server/src/main/resources/ServerResourceBundle_be.properties
#	server/src/main/resources/ServerResourceBundle_pl.properties
#	server/src/main/resources/ServerResourceBundle_ru.properties
#	server/src/main/resources/ServerResourceBundle_uk.properties
#	server/src/main/resources/ServerResourceBundle_uz.properties
#	server/src/main/resources/ServerResourceBundle_zh.properties

+++ server/src/main/java/lsfusion/server/data/query/exec/materialize/MaterializedQuery.java
@@ -11,7 +11,7 @@ import lsfusion.server.data.type.exec.EnsureTypeEnvironment;
 public class MaterializedQuery {
     public final String tableName;
     public final String mapFields;
-    public final int count;
+    public final long count;
     public final long timeExec;
 
     // for debug
@@ -31,7 +31,7 @@ public class MaterializedQuery {
         }
     }
 
-    public MaterializedQuery(String tableName, String mapFields, ImOrderSet<KeyField> keyFields, ImSet<PropertyField> propFields, int count, long timeExec, Owner owner) {
+    public MaterializedQuery(String tableName, String mapFields, ImOrderSet<KeyField> keyFields, ImSet<PropertyField> propFields, long count, long timeExec, Owner owner) {
         this.tableName = tableName;
         this.mapFields = mapFields;
         this.count = count;
