+++ server/src/main/java/lsfusion/server/data/table/FillTemporaryTable.java
@@ -6,7 +6,7 @@ import java.sql.SQLException;
 
 public abstract class FillTemporaryTable {
 
-    public abstract Integer fill(String name) throws SQLException, SQLHandledException;
+    public abstract Long fill(String name) throws SQLException, SQLHandledException;
 
     public boolean canBeNotEmptyIfFailed() {
         return false;
